Extracted common IO class code into a superclass.
